requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.0
charset-normalizer==3.2.0
cryptography==3.0
defusedxml==0.7.1
Django==4.2.2
django-cors-headers==4.2.0
django-froala-editor==4.1.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-queryfields==1.0.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
idna==3.4
oauthlib==3.2.2
Pillow==8.0.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.6
